Internal refactor
